+++ src/app/app.module.ts
@@ -11,27 +11,26 @@ import {AuthorizationHttpInterceptorService} from './services/authorizationhttpi
 import {JsonDateInterceptorService} from './services/json-date-interceptor.service';
 
 @NgModule({
-  declarations: [AppComponent],
-  entryComponents: [],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    IonicModule.forRoot(),
-    AppRoutingModule
-  ],
-  providers: [
-    {provide: RouteReuseStrategy, useClass: IonicRouteStrategy},
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: AuthorizationHttpInterceptorService,
-      multi: true,
-    },
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: JsonDateInterceptorService,
-      multi: true
-    },
-  ],
-  bootstrap: [AppComponent]
+    declarations: [AppComponent],
+    imports: [
+        BrowserModule,
+        HttpClientModule,
+        IonicModule.forRoot(),
+        AppRoutingModule
+    ],
+    providers: [
+        { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: AuthorizationHttpInterceptorService,
+            multi: true,
+        },
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: JsonDateInterceptorService,
+            multi: true
+        },
+    ],
+    bootstrap: [AppComponent]
 })
 export class AppModule {}