+++ src/app/models/hourlypoweraverage.ts
@@ -1,0 +1,6 @@
+export class HourlyPowerAverage {
+    public hour: Date;
+    public averagePowerUsage: number;
+    public averagePowerProduction: number;
+    public averageGasFlow: number;
+}