+++ src/app/pages/energy/monthly/monthly.module.ts
@@ -10,7 +10,7 @@ import { MonthlyPage } from './monthly.page';
 import {CostChartModule} from '../../../components/cost-chart/cost-chart.module';
 import {EnergyBarChartModule} from '../../../components/energy-bar-chart/energy-bar-chart.module';
 import {InfoCardModule} from '../../../components/info-card/info-card.module';
-import {GasChartModule} from "../../../components/gas-chart/gas-chart.module";
+import {GasChartModule} from '../../../components/gas-chart/gas-chart.module';
 
 @NgModule({
     imports: [