+++ src/app/pages/energy/monthly/monthly.page.ts
@@ -6,6 +6,8 @@ import {mergeMap} from 'rxjs/operators';
 import {HistoricData} from '../../../models/historicdata';
 import {EnergyUsage} from '../../../models/energyusage';
 import {Device} from '../../../models/device';
+import {Response} from '../../../models/response';
+import { CostCalculatorService } from 'app/services/cost-calculator.service';
 
 @Component({
   selector: 'app-monthly',
@@ -19,6 +21,8 @@ export class MonthlyPage implements OnInit, AfterViewInit {
   public powerUsage: string;
   public powerProduction: string;
   public netEnergyUsage: string;
+  public energyUsagePerDay: string;
+  public averageCostPerDay: string;
   public costLabels: string[];
   public costValues: number[];
   public barChartPowerUsage: number[];
@@ -32,6 +36,7 @@ export class MonthlyPage implements OnInit, AfterViewInit {
     'Jun', 'Jul', 'Aug', 'Sept', 'Oct', 'Nov', 'Dec'];
 
   public constructor(private readonly dsmr: DsmrService,
+                     private readonly costCalculator: CostCalculatorService,
                      private readonly settings: SettingsService) {
     this.barGasUsage = [];
     this.labels = [];
@@ -65,21 +70,21 @@ export class MonthlyPage implements OnInit, AfterViewInit {
   }
 
   private loadData() {
-    return new Promise((resolve, reject) => {
+    return new Promise<void>((resolve, reject) => {
       const startDate = MonthlyPage.getFirstOfMonth();
       const endDate = MonthlyPage.getEndToday();
       const device = this.dsmr.getSelectedDevice();
 
       this.dsmr.getPowerData(device.id, startDate, endDate, 'day')
-        .pipe(mergeMap(result => {
+        .pipe(mergeMap((result: Response<EnergyDataPoint[]>) => {
           this.computeCostChart(result.data);
           return this.dsmr.getEnergyUsage(device.id, startDate, endDate);
-        }),mergeMap(result => {
+        }),mergeMap((result: Response<EnergyUsage>) => {
           this.computeCards(result.data);
           const firstMonthDate = MonthlyPage.addMonths(startDate, -3);
 
           return this.dsmr.getPowerData(device.id, firstMonthDate, endDate, 'day');
-        })).subscribe(result => {
+        })).subscribe((result: Response<EnergyDataPoint[]>) => {
           this.computeCharts(result.data);
           resolve();
       }, _ => {
@@ -122,44 +127,44 @@ export class MonthlyPage implements OnInit, AfterViewInit {
 
     this.labels = labels;
     this.barGasUsage = gasUsage;
-    this.barChartPowerProduction = powerProduction;
+
+    if(this.device.hasSolarCells) {
+      this.barChartPowerProduction = powerProduction;
+    }
+
     this.barChartPowerUsage = powerUsage;
   }
 
   private computeCards(data: EnergyUsage) {
     const netUsage = data.energyUsage - data.energyProduction;
+    const today = new Date();
+
     this.netEnergyUsage = netUsage.toFixed(2);
     this.powerUsage = data.energyUsage.toFixed(2);
     this.powerProduction = data.energyProduction.toFixed(2);
     this.gasUsageMonthly = data.gasUsage?.toFixed(2);
-    this.cost = this.computeCost(data).toFixed(2);
-  }
-
-  private computeCost(usage: EnergyUsage) {
-    const prices = this.settings.getPrices();
-    let cost = 0;
-
-    cost += prices.powerUsage * usage.energyUsage;
-    cost += prices.gas * usage.gasUsage;
-    cost -= prices.powerProduction * usage.energyProduction;
-
-    return cost;
+    this.cost = this.costCalculator.computeCostForEnergyUsage(data).toFixed(2);
+    this.energyUsagePerDay = (netUsage / today.getDate()).toFixed(2);
   }
 
   private computeCostChart(data: EnergyDataPoint[]) {
     const prices = this.settings.getPrices();
     const labels: string[] = [];
     const values: number[] = [];
+    let average = 0;
 
     data.forEach(x => {
       let cost = x.energyUsage / 1000 * prices.powerUsage;
       cost -= x.energyProduction / 1000 * prices.powerProduction;
       cost += x.gasFlow * prices.gas;
 
+      average += cost;
       values.push(cost);
       labels.push(`${MonthlyPage.padNumer(x.timestamp.getDate(), 2)}-${MonthlyPage.padNumer(x.timestamp.getMonth(), 2)}`);
     });
 
+    average /= data.length;
+    this.averageCostPerDay = average.toFixed(2);
     this.costLabels = labels;
     this.costValues = values;
   }