+++ src/app/pages/energy/weekly/weekly.page.ts
@@ -3,9 +3,10 @@ import {DsmrService} from '../../../services/dsmr.service';
 import { Chart } from 'chart.js';
 import {EnergyDataPoint} from '../../../models/energydatapoint';
 import {SettingsService} from '../../../services/settings.service';
+import {Response} from '../../../models/response';
 import {mergeMap} from 'rxjs/operators';
-import {GroupedPowerData} from '../../../models/groupedpowerdata';
 import {Device} from '../../../models/device';
+import { HourlyPowerAverage } from 'app/models/HourlyPowerAverage';
 
 @Component({
   selector: 'app-weekly',
@@ -18,7 +19,8 @@ export class WeeklyPage implements OnInit, AfterViewInit {
   public costLabels: string[];
   public costValues: number[];
   public gasUsageToday: string;
-  public credit: string;
+  public averageDailyEnergyUsage: string;
+  public averageDailyCost: string;
   public cost: string;
   public powerUsage: string;
   public powerProduction: string;
@@ -90,20 +92,22 @@ export class WeeklyPage implements OnInit, AfterViewInit {
   }
 
   private loadData() {
-    return new Promise((resolve, reject) => {
+    return new Promise<void>((resolve, reject) => {
       const startDate = WeeklyPage.getStartLastWeek();
       const endDate = WeeklyPage.getEndToday();
       const device = this.dsmr.getSelectedDevice();
 
-      this.dsmr.getPowerData(device.id, startDate, endDate, 'day').pipe(mergeMap(result => {
+      this.dsmr.getPowerData(device.id, startDate, endDate, 'day').pipe(mergeMap((result: Response<EnergyDataPoint[]>) => {
         this.computeCards(result.data);
         this.computeCharts(result.data);
         this.computeCostChart(result.data);
 
-        return this.dsmr.getGroupedPowerData(device.id);
-      })).subscribe(result => {
-        this.computeGroupedChart(result.data);
-
+        return this.dsmr.getAverageEnergyData(device.id, WeeklyPage.getStartLastWeek(), WeeklyPage.getEndToday());
+      }), mergeMap((result: Response<HourlyPowerAverage[]>) => {
+        this.computeEnergyAveragesThisWeek(result.data);
+        return this.dsmr.getAverageEnergyData(device.id, WeeklyPage.getStartLastMonth(), WeeklyPage.getEndLastWeek());
+      })).subscribe((result: Response<HourlyPowerAverage[]>) => {
+        this.computeEnergyAveragesThisMonth(result.data);
         this.refreshView();
         resolve();
       }, _ => {
@@ -112,26 +116,35 @@ export class WeeklyPage implements OnInit, AfterViewInit {
     });
   }
 
-  private computeGroupedChart(data: GroupedPowerData[]) {
-    const usage: number [] = [];
-    const production: number[] = [];
-    const labels: string[] = [];
+  private computeEnergyAveragesThisMonth(data: HourlyPowerAverage[]) {
+    const resultsLastMonth: number[] = [];
 
     data.forEach(x => {
-      if (x.hour < 6 || x.hour > 22) {
+      if(x.hour.getUTCHours() < 6 || x.hour.getUTCHours() > 23) {
         return;
       }
 
-      production.push(x.production / 1000);
-      usage.push(x.usage / 1000);
+      resultsLastMonth.push(x.averagePowerUsage);
+    });
 
-      const hour = WeeklyPage.padNumer(x.hour, 2);
-      labels.push(`${hour}:00`);
+    this.groupedEnergyProduction = resultsLastMonth;
+  }
+
+  private computeEnergyAveragesThisWeek(data: HourlyPowerAverage[]) {
+    const resultsThisWeek: number[] = [];
+    const labels: string[] = [];
+
+    data.forEach(x => {
+      if(x.hour.getUTCHours() < 6 || x.hour.getUTCHours() > 23) {
+        return;
+      }
+
+      resultsThisWeek.push(x.averagePowerUsage);
+      labels.push(`${x.hour.getUTCHours()}:00`);
     });
 
     this.groupedLabels = labels;
-    this.groupedEnergyProduction = production;
-    this.groupedEnergyUsage = usage;
+    this.groupedEnergyUsage = resultsThisWeek;
   }
 
   private computeCharts(data: EnergyDataPoint[]) {
@@ -140,7 +153,10 @@ export class WeeklyPage implements OnInit, AfterViewInit {
     const labels: string[] = [];
 
     data.forEach(dp => {
-      production.push(dp.energyProduction / 1000);
+      if(this.device.hasSolarCells) {
+        production.push(dp.energyProduction / 1000);
+      }
+
       usage.push(dp.energyUsage / 1000);
       this.lineGasUsage.push(dp.gasFlow);
 
@@ -170,10 +186,15 @@ export class WeeklyPage implements OnInit, AfterViewInit {
     this.powerUsage = usage.toFixed(2);
     this.powerProduction = production.toFixed(2);
     this.gasUsageToday = gasUsage.toFixed(2);
-    this.cost = this.computeCost(usage, production, gasUsage).toFixed(2);
+    this.averageDailyEnergyUsage = (usage / data.length).toFixed(2);
+
+    this.computeCostValues(usage, production, gasUsage, data.length);
   }
 
-  private computeTariff() {
+  private computeCostValues(usage: number, production: number, gas: number, count: number) {
+    const cost = this.computeCost(usage, production, gas);
+    this.cost = cost.toFixed(2);
+    this.averageDailyCost = (cost / count).toFixed(2);
   }
 
   private computeCost(usage: number, production: number, gas: number) {
@@ -269,6 +290,27 @@ export class WeeklyPage implements OnInit, AfterViewInit {
     todayEnd.setMilliseconds(999);
 
     return todayEnd;
+  }
+
+    // eslint-disable-next-line @typescript-eslint/member-ordering
+  private static getEndLastWeek() {
+    const endLastWeek = new Date();
+
+    endLastWeek.setHours(23, 59, 59);
+    endLastWeek.setMilliseconds(0);
+    endLastWeek.setDate(endLastWeek.getDate() - 7);
+    return endLastWeek;
+  }
+
+    // eslint-disable-next-line @typescript-eslint/member-ordering
+  private static getStartLastMonth() {
+    const lastMonth = new Date();
+
+    lastMonth.setHours(0,0,0);
+    lastMonth.setMilliseconds(0);
+    lastMonth.setMonth(lastMonth.getMonth() - 1);
+    lastMonth.setDate(lastMonth.getDate() - 7);
+    return lastMonth;
   }
 
   // eslint-disable-next-line @typescript-eslint/member-ordering