+++ src/app/pages/reports/reports.page.ts
@@ -6,7 +6,10 @@ import {EnergyDataPoint} from '../../models/energydatapoint';
 import {HistoricData} from '../../models/historicdata';
 import {GroupedPowerData} from '../../models/groupedpowerdata';
 import {EnergyUsage} from '../../models/energyusage';
-import {Device} from "../../models/device";
+import {Device} from '../../models/device';
+import {Response} from '../../models/response';
+import { stringify } from 'querystring';
+import { CostCalculatorService } from 'app/services/cost-calculator.service';
 
 @Component({
   selector: 'app-reports',
@@ -19,6 +22,7 @@ export class ReportsPage implements OnInit, AfterViewInit {
   public powerUsage: string;
   public netEnergyUsage: string;
   public powerProduction: string;
+  public averageCostPerDay: string;
   public costLabels: string[];
   public costValues: number[];
   public barChartPowerUsage: number[];
@@ -42,6 +46,7 @@ export class ReportsPage implements OnInit, AfterViewInit {
     'Jun', 'Jul', 'Aug', 'Sept', 'Oct', 'Nov', 'Dec'];
 
   public constructor(private readonly dsmr: DsmrService,
+                     private readonly costCalculator: CostCalculatorService,
                      private readonly settings: SettingsService) {
     this.barGasUsage = [];
     this.labels = [];
@@ -92,23 +97,23 @@ export class ReportsPage implements OnInit, AfterViewInit {
   }
 
   private loadData() {
-    return new Promise((resolve, reject) => {
+    return new Promise<void>((resolve, reject) => {
       const startDate = this.getFirstOfMonth();
       const endDate = this.getEndOfMonth();
       const device = this.dsmr.getSelectedDevice();
 
-      this.dsmr.getPowerData(device.id, startDate, endDate, 'day').pipe(mergeMap(result => {
+      this.dsmr.getPowerData(device.id, startDate, endDate, 'day').pipe(mergeMap((result: Response<EnergyDataPoint[]>) => {
         this.computeCostChart(result.data);
         return this.dsmr.getEnergyUsage(device.id, startDate, endDate);
-      }), mergeMap(result => {
+      }), mergeMap((result: Response<EnergyUsage>) => {
         this.computeCards(result.data);
         return this.dsmr.getGroupedPowerDataBetween(device.id, startDate, endDate);
-      }), mergeMap(result => {
+      }), mergeMap((result: Response<GroupedPowerData[]>) => {
         this.computeGroupedChart(result.data);
 
         const firstMonthDate = ReportsPage.addMonths(startDate, -3);
         return this.dsmr.getPowerData(device.id, firstMonthDate, endDate, 'day');
-      })).subscribe(result => {
+      })).subscribe((result: Response<EnergyDataPoint[]>) => {
         this.computeOverviewCharts(result.data);
 
         resolve();
@@ -150,7 +155,10 @@ export class ReportsPage implements OnInit, AfterViewInit {
       labels.push(ReportsPage.months[key]);
     });
 
-    this.barChartPowerProduction = powerProduction;
+    if(this.device.hasSolarCells) {
+      this.barChartPowerProduction = powerProduction;
+    }
+
     this.barChartPowerUsage = powerUsage;
     this.barGasUsage = gasUsage;
     this.labels = labels;
@@ -167,7 +175,7 @@ export class ReportsPage implements OnInit, AfterViewInit {
     this.powerUsage = data.energyUsage.toFixed(2);
     this.powerProduction = data.energyProduction.toFixed(2);
     this.gasUsageMonthly = data.gasUsage?.toFixed(2);
-    this.cost = this.computeCost(data).toFixed(2);
+    this.cost = this.costCalculator.computeCostForEnergyUsage(data).toFixed(2);
   }
 
   private setNullCards() {
@@ -176,33 +184,27 @@ export class ReportsPage implements OnInit, AfterViewInit {
     this.powerProduction = '0';
     this.gasUsageMonthly = '0';
     this.cost = '0';
-  }
-
-  private computeCost(usage: EnergyUsage) {
-    const prices = this.settings.getPrices();
-    let cost = 0;
-
-    cost += prices.powerUsage * usage.energyUsage;
-    cost += prices.gas * usage.gasUsage;
-    cost -= prices.powerProduction * usage.energyProduction;
-
-    return cost;
+    this.averageCostPerDay = '0';
   }
 
   private computeCostChart(data: EnergyDataPoint[]) {
     const prices = this.settings.getPrices();
     const labels: string[] = [];
     const values: number[] = [];
+    let average = 0;
 
     data.forEach(x => {
       let cost = x.energyUsage / 1000 * prices.powerUsage;
       cost -= x.energyProduction / 1000 * prices.powerProduction;
       cost += x.gasFlow * prices.gas;
+      average += cost;
 
       values.push(cost);
       labels.push(`${ReportsPage.padNumer(x.timestamp.getDate(), 2)}-${ReportsPage.padNumer(x.timestamp.getMonth() + 1, 2)}`);
     });
 
+    average /= data.length;
+    this.averageCostPerDay = average.toFixed(2);
     this.costLabels = labels;
     this.costValues = values;
   }
@@ -225,7 +227,11 @@ export class ReportsPage implements OnInit, AfterViewInit {
     });
 
     this.groupedLabels = labels;
-    this.groupedEnergyProduction = production;
+
+    if(this.device.hasSolarCells) {
+      this.groupedEnergyProduction = production;
+    }
+
     this.groupedEnergyUsage = usage;
   }