+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 23-12-2021
+### Added
+- Additional aggregates
+- Solar cell detection
+
+### Updated
+- Views for devices without solar cells
+- Energy by hour chart on the overview page
+
+## [1.0.0] - 08-08-2021
 ### Added
 - Power pages
 - Current situation view